port r74811 from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-3-5-branch@74812 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/radixsort.c
@@ -140,7 +140,7 @@ static void growstack(uint64_t newlen)
     if (gs[flip] == NULL)
 	Error("Failed to realloc working memory stack to %d*4bytes (flip=%d)",
 	      (int)newlen /* no bigger than gsmaxalloc */, flip);
-    gsalloc[flip] = newlen;
+    gsalloc[flip] = (int)newlen;
 }
 
 static void push(int x)
